test(gemma4): split the prefill-config and norm-sharding tests out of test_ccl_topology
test_ccl_topology.py had grown to 630 lines / 41 cases while its docstring still
claimed "CCL topology / async / L1 env knobs". Most of what arrived after
cea54a0f9f4 was not CCL at all: tuned prefill matmul program configs, the o_proj
opt-in, RMSNorm width-sharding and the prefill LN island. Moved those out so each
file's name says what is in it.

  test_prefill_matmul_configs.py  8 tests over tt/dram_sharded.py program configs
                                  (+ the _MLP_PREFILL_SHAPES table and the
                                  _mlp_prefill_config dispatch they share)
  test_norm_sharding.py           3 tests over tt/rms_norm.py shard specs and the
                                  island kill switches
  host_config_fakes.py            _FakeComputeGrid / _FakeMeshDevice / _FakeTensor,
                                  which both the moved and the remaining tests
                                  construct

Scoped to post-cea54a0f9f4 code, which is why test_ccl_topology.py keeps a few
older non-CCL cases (test_dram_shard_*, test_prefill_progcfg_in0_block_w_divides_kt,
test_weight_cache_path_*): those predate the baseline. Its docstring is left alone
for the same reason -- the split makes it more accurate, not less.

Nothing was rewritten. Every test body, parametrize table, and fake class moved
verbatim; the only new lines are the three module headers and the
host_config_fakes imports.

Verified: definition-by-definition AST comparison across the four files shows 40
definitions before and 40 after, with none lost, added, changed, or duplicated.
41 tests collected at HEAD, 41 pass after the split -- run under both real
checkpoints, 12B (/home/user/gemma-4-12B-it) and 31B (google/gemma-4-31B-it) at
MESH_DEVICE=1x8, 41 passed each. These are host-only cases; no device is opened.
pyflakes clean on all four files, pre-commit clean twice. No CI yaml names
test_ccl_topology.py, and none of the new test names match the unit job's -k
filters, so the split changes nothing about what CI collects.
